TIDIJ-191 Moved ui elements to core
