+++ core/lib/ui_elements/vehicle_controller.dart
@@ -1,5 +1,5 @@
 import 'package:fleetcarpooling/VehicleManagamentService/vehicle_managament_service.dart';
-import 'package:fleetcarpooling/ui_elements/custom_toast.dart';
+import 'package:core/ui_elements/custom_toast.dart';
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'colors' as my_defined_colors;
