+++ fleetcarpooling/lib/chat/widgets/custom_text_form_field.dart
@@ -1,4 +1,4 @@
-import 'package:fleetcarpooling/ui_elements/colors';
+import 'package:core/ui_elements/colors';
 import 'package:flutter/material.dart';
 
 class CustomTextFormField extends StatelessWidget {
