+++ fleetcarpooling/lib/pages/reset_password_form.dart
@@ -1,8 +1,9 @@
+import 'package:core/ui_elements/buttons.dart';
 import 'package:fleetcarpooling/auth/user_repository.dart';
-import 'package:fleetcarpooling/ui_elements/buttons.dart';
-import 'package:fleetcarpooling/ui_elements/colors';
+
+import 'package:core/ui_elements/colors';
 import 'package:flutter/material.dart';
-import 'package:fleetcarpooling/ui_elements/text_field.dart';
+import 'package:core/ui_elements/text_field.dart';
 
 class ResetPasswordForm extends StatefulWidget {
   const ResetPasswordForm({super.key});
