+++ qraddition/lib/add_vehicle_QR_form.dart
@@ -3,9 +3,9 @@ import 'package:core/add_vehicle_interface.dart';
 import 'package:core/bloc/vehicle_bloc.dart';
 import 'package:core/event/vehicle_event.dart';
 import 'package:core/state/vehicle_state.dart';
+import 'package:core/ui_elements/buttons.dart';
+import 'package:core/ui_elements/colors';
 import 'package:core/vehicle.dart';
-import 'package:fleetcarpooling/ui_elements/buttons.dart';
-import 'package:fleetcarpooling/ui_elements/colors';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:qr_mobile_vision/qr_camera.dart';
